Ruff

+++ tests/test_data_retrieval.py
@@ -1,38 +1,34 @@
+"""Test data retrieval."""
+from __future__ import annotations
+
+import shutil
 import unittest
-from pathlib import Path
 from datetime import datetime
+from pathlib import Path
 
 import pandas as pd
-import shutil
+from sentinelhub import DataCollection, Geometry, SentinelHubStatistical
 
-from sat4ec.data_retrieval import IndicatorData as IData
-from sat4ec.system.collections import SubsetCollection as Subsets
 from sat4ec.aoi_check import AOI
-from sat4ec.system.helper_functions import get_last_month
+from sat4ec.data_retrieval import IndicatorData as IData
 from sat4ec.main import run_indicator
-
-from sentinelhub import (
-    Geometry,
-    DataCollection,
-    SentinelHubStatistical,
-)
-
+from sat4ec.system.helper_functions import get_last_month
+from sat4ec.system.subset_collections import SubsetCollection as Subsets
 
 TEST_DIR = Path(r"/mnt/data1/gitlab/sat4ec/tests/testdata")
 
 
 class TestGetData(unittest.TestCase):
-    def __init__(self, *args, **kwargs):
-        super(TestGetData, self).__init__(*args, **kwargs)
+    """Encapsulates testing methods."""
+
+    def __init__(self: TestGetData, *args: int, **kwargs: int) -> None:
+        """Initialize TestGetData class."""
+        super().__init__(*args, **kwargs)
         self.tear_down = True  # delete output data per default, switch to False in test methods if required
         self.out_dir = TEST_DIR.joinpath("output", "vw_wolfsburg")
-        self.daily_out_file = TEST_DIR.joinpath(
-            "orbit_input", "raw", "indicator_1_rawdata_daily_aoi_split_asc_VH.csv"
-        )
-        self.aoi_collection = AOI(
-            data=TEST_DIR.joinpath("input", "AOIs", "vw_wolfsburg_aoi_split.geojson")
-        )
-        self.features = [feature for feature in self.aoi_collection.get_feature()]
+        self.daily_out_file = TEST_DIR.joinpath("orbit_input", "raw", "indicator_1_rawdata_daily_aoi_split_asc_VH.csv")
+        self.aoi_collection = AOI(data=TEST_DIR.joinpath("input", "AOIs", "vw_wolfsburg_aoi_split.geojson"))
+        self.features = list(self.aoi_collection.get_feature())
         self.feature = self.features[0]
 
         self.indicator = IData(
@@ -52,19 +48,21 @@ class TestGetData(unittest.TestCase):
             pol=self.indicator.pol,
         )
 
-    def tearDown(self):
-        if self.tear_down:
-            if self.out_dir.exists():
-                shutil.rmtree(self.out_dir)
+    def tearDown(self: TestGetData) -> None:
+        """Delete test output data."""
+        if self.tear_down and self.out_dir.exists():
+            shutil.rmtree(self.out_dir)
 
-    def test_class_init(self):
+    def test_class_init(self: TestGetData) -> None:
+        """Test initialization of class."""
         self.assertTrue(isinstance(self.indicator.geometry, Geometry))
         self.assertTrue(isinstance(self.indicator.size, tuple))
         self.assertEqual(len(self.indicator.size), 2)
         self.assertTrue(isinstance(self.indicator.collection, DataCollection))
         self.assertTrue(self.indicator.out_dir.exists())
 
-    def test_no_dates_provided(self):
+    def test_no_dates_provided(self: TestGetData) -> None:
+        """Test behavior if no dates were provided."""
         indicator = IData(
             aoi=self.feature.geometry,
             fid=self.feature.fid,
@@ -74,7 +72,8 @@ class TestGetData(unittest.TestCase):
         self.assertEqual(indicator.start_date, "2014-05-01")
         self.assertEqual(indicator.end_date, get_last_month())
 
-    def test_existing_data_new_past_dates(self):
+    def test_existing_data_new_past_dates(self: TestGetData) -> None:
+        """Test archive data with past data set."""
         self.subsets.daily_out_file = self.daily_out_file
         self.subsets.check_existing_raw()
 
@@ -91,18 +90,15 @@ class TestGetData(unittest.TestCase):
         )
 
         self.assertTrue(indicator.check_dates(start=True) == "past")
-        self.assertEqual(
-            datetime.strftime(indicator.end_date, "%Y-%m-%d"), "2020-01-03"
-        )
+        self.assertEqual(datetime.strftime(indicator.end_date, "%Y-%m-%d"), "2020-01-03")
 
         run_indicator(indicator)
         past_df = indicator.dataframe
         indicator.concat_dataframes(past_df=past_df)
-        self.assertEqual(
-            len(indicator.dataframe), len(self.subsets.archive_dataframe) + 4
-        )
+        self.assertEqual(len(indicator.dataframe), len(self.subsets.archive_dataframe) + 4)
 
-    def test_existing_data_new_future_dates(self):
+    def test_existing_data_new_future_dates(self: TestGetData) -> None:
+        """Test archive data with future data set."""
         self.subsets.daily_out_file = self.daily_out_file
         self.subsets.check_existing_raw()
 
@@ -119,22 +115,17 @@ class TestGetData(unittest.TestCase):
         )
 
         self.assertTrue(indicator.check_dates(end=True) == "future")
-        self.assertEqual(
-            datetime.strftime(indicator.start_date, "%Y-%m-%d"), "2023-09-27"
-        )
+        self.assertEqual(datetime.strftime(indicator.start_date, "%Y-%m-%d"), "2023-09-27")
 
         run_indicator(indicator)
         future_df = indicator.dataframe
         indicator.concat_dataframes(future_df=future_df)
-        self.assertEqual(
-            len(indicator.dataframe), len(self.subsets.archive_dataframe) + 5
-        )
+        self.assertEqual(len(indicator.dataframe), len(self.subsets.archive_dataframe) + 5)
 
-    def test_existing_data_non_existing_column(self):
+    def test_existing_data_non_existing_column(self: TestGetData) -> None:
+        """Test if archive data exists but column does not exist."""
         self.subsets.archive_dataframe = pd.read_csv(self.daily_out_file)
-        self.subsets.archive_dataframe = self.subsets.archive_dataframe.set_index(
-            "interval_from"
-        )
+        self.subsets.archive_dataframe = self.subsets.archive_dataframe.set_index("interval_from")
         self.indicator.archive_data = self.subsets.archive_dataframe
         self.indicator.fid = "11"
         existing_keyword, column_keyword = self.indicator.check_existing_data()
@@ -142,45 +133,40 @@ class TestGetData(unittest.TestCase):
         self.assertTrue(existing_keyword)
         self.assertFalse(column_keyword)
 
-    def test_non_existing_data(self):
+    def test_non_existing_data(self: TestGetData) -> None:
+        """Test if archive exists on non-existing data."""
         existing_keyword, column_keyword = self.indicator.check_existing_data()
 
         self.assertFalse(existing_keyword)
         self.assertIsNone(column_keyword)
 
-    def test_request_grd(self):
+    def test_request_grd(self: TestGetData) -> None:
+        """Test Sentinel Hub request."""
         self.indicator.get_request_grd()
         self.assertTrue(isinstance(self.indicator.request, SentinelHubStatistical))
 
-    def test_get_data(self):
+    def test_get_data(self: TestGetData) -> None:
+        """Test data retrieval from Sentinel Hub."""
         self.indicator.get_request_grd()
         self.indicator.get_data()
         self.assertEqual(list(self.indicator.stats.keys()), ["data", "status"])
         self.assertEqual(self.indicator.stats["status"], "OK")
 
-    def test_stats_to_df(self):
+    def test_stats_to_df(self: TestGetData) -> None:
+        """Test translating statistics into dataframe."""
         self.indicator.get_request_grd()
         self.indicator.get_data()
         self.indicator.stats_to_df()
 
-        self.assertTrue(
-            self.indicator.dataframe.dtypes["interval_to"], pd.DatetimeTZDtype
-        )
-        self.assertTrue(
-            self.indicator.dataframe.dtypes[f"{self.feature.fid}_min"], "float32"
-        )
-        self.assertTrue(
-            self.indicator.dataframe.dtypes[f"{self.feature.fid}_max"], "float32"
-        )
-        self.assertTrue(
-            self.indicator.dataframe.dtypes[f"{self.feature.fid}_mean"], "float32"
-        )
-        self.assertTrue(
-            self.indicator.dataframe.dtypes[f"{self.feature.fid}_std"], "float32"
-        )
+        self.assertTrue(self.indicator.dataframe.dtypes["interval_to"], pd.DatetimeTZDtype)
+        self.assertTrue(self.indicator.dataframe.dtypes[f"{self.feature.fid}_min"], "float32")
+        self.assertTrue(self.indicator.dataframe.dtypes[f"{self.feature.fid}_max"], "float32")
+        self.assertTrue(self.indicator.dataframe.dtypes[f"{self.feature.fid}_mean"], "float32")
+        self.assertTrue(self.indicator.dataframe.dtypes[f"{self.feature.fid}_std"], "float32")
         self.assertTrue(self.indicator.dataframe.index.inferred_type, pd.DatetimeIndex)
 
-    def test_monthly_aggregate(self):
+    def test_monthly_aggregate(self: TestGetData) -> None:
+        """Test aggregation of semi-daily to monthly data."""
         self.indicator.monthly = True
         self.indicator.get_request_grd()
         self.indicator.get_data()
@@ -194,13 +180,10 @@ class TestGetData(unittest.TestCase):
         self.assertTrue(self.subsets.dataframe.index.inferred_type, pd.DatetimeIndex)
         self.assertTrue(len(self.subsets.dataframe) < len(daily_dataframe))
 
-    def test_regression_raw(self):
-        self.indicator.dataframe = pd.read_csv(
-            self.out_dir.joinpath("raw", "indicator_1_rawdata_asc_VH.csv")
-        )
-        self.indicator.dataframe["interval_from"] = pd.to_datetime(
-            self.indicator.dataframe["interval_from"]
-        )
+    def test_regression_raw(self: TestGetData) -> None:
+        """Test regression on raw data."""
+        self.indicator.dataframe = pd.read_csv(self.out_dir.joinpath("raw", "indicator_1_rawdata_asc_VH.csv"))
+        self.indicator.dataframe["interval_from"] = pd.to_datetime(self.indicator.dataframe["interval_from"])
         self.indicator.dataframe = self.indicator.dataframe.set_index("interval_from")
         self.subsets.dataframe = self.indicator.dataframe
         self.subsets.features = self.features
@@ -221,13 +204,10 @@ class TestGetData(unittest.TestCase):
             ).exists()
         )
 
-    def test_regression_monthly(self):
-        self.indicator.dataframe = pd.read_csv(
-            self.out_dir.joinpath("raw", "indicator_1_rawdata_asc_VH.csv")
-        )
-        self.indicator.dataframe["interval_from"] = pd.to_datetime(
-            self.indicator.dataframe["interval_from"]
-        )
+    def test_regression_monthly(self: TestGetData) -> None:
+        """Test regression on monthly data."""
+        self.indicator.dataframe = pd.read_csv(self.out_dir.joinpath("raw", "indicator_1_rawdata_asc_VH.csv"))
+        self.indicator.dataframe["interval_from"] = pd.to_datetime(self.indicator.dataframe["interval_from"])
         self.indicator.dataframe = self.indicator.dataframe.set_index("interval_from")
         self.subsets.dataframe = self.indicator.dataframe
         self.subsets.features = self.features
@@ -250,13 +230,14 @@ class TestGetData(unittest.TestCase):
             ).exists()
         )
 
-    def test_save_df_raw(self):  # pure raw data
+    def test_save_df_raw(self: TestGetData) -> None:  # pure raw data
+        """Test saving semi-daily data."""
         self.tear_down = False
         self.indicator.get_request_grd()
         self.indicator.get_data()
         self.indicator.stats_to_df()
         self.subsets.dataframe = self.indicator.dataframe
-        self.subsets.save_raw()
+        self.subsets.save_daily_raw()
 
         self.assertTrue(
             self.indicator.out_dir.joinpath(
@@ -265,14 +246,15 @@ class TestGetData(unittest.TestCase):
             ).exists()
         )
 
-    def test_save_df_monthly(self):  # monthly raw data
+    def test_save_df_monthly(self: TestGetData) -> None:  # monthly raw data
+        """Test saving monthly data."""
         self.indicator.get_request_grd()
         self.indicator.get_data()
         self.indicator.stats_to_df()
         self.subsets.dataframe = self.indicator.dataframe
         self.subsets.monthly = True
         self.subsets.monthly_aggregate()
-        self.subsets.save_raw()
+        self.subsets.save_monthly_raw()
 
         self.assertTrue(
             self.indicator.out_dir.joinpath(
